refactored the filter by both button to instead be a checkbox that sets the filter by both state to instantly filter by both options when checked; now users do not have to click a button every time they make a change.
