+++ api/notes.json
@@ -1,8 +1,8 @@
 {
   "notes": [
     {
-      "date": "2020-12-09",
-      "noteText": "Maybe...with a name like, Hilario.....",
+      "date": "2020-12-08",
+      "noteText": "Hmmm, this guy looks fishy. Like, really, really fishy.",
       "criminalId": 78,
       "id": 1
     },
@@ -29,6 +29,12 @@
       "noteText": "Ezekiel is a cool name. This is the final test....I think.",
       "criminalId": 98,
       "id": 5
+    },
+    {
+      "date": "2021-01-05",
+      "noteText": "No, THIS is the last test...until the nest last text. 🙄 He still looks fishy.",
+      "criminalId": 78,
+      "id": 6
     }
   ]
 }
