+++ index.html
@@ -12,9 +12,9 @@
     <main id="container">
         <header id="headerContainer">
             <div id="filters">
+                <div id="filterCheckbox"></div>
                 <div class="filter" id="convictionFilter"></div>
                 <div class="filter" id="officerFilter"></div>
-                <div id="filterButton"></div>
                 <div id="themeRadios"></div>
             </div>
             <div id="buttonContainer"></div>
