+++ scripts/criminals/CriminalConvictionKey.js
@@ -3,8 +3,7 @@
 *   of HTML giving structure to a color key that allows users to easily identify the conviction type.
 */
 
-import { switchFunction } from "./Criminal.js";
-
+import { switchFunction } from "../utils/switchFunction.js";
 
 // Function, criminal, that accepts an argument, criminalObject, and returns a string of an HTML element.
 export const CriminalConvictionKey = (conviction) => {
