+++ scripts/officers/officersDataProvider.js
@@ -3,22 +3,14 @@
  *   the array officers, and then copies the array and exports it for use elsewhere with, useOfficers.
  */
 
+import { sortArrayByLastName } from "../utils/sortArrayByLastName.js";
+
 // Set empty array for getOfficers to place the parsed data in.
 let officers = [];
 
 // Returns a copy of the array officers to be used later.
 export const useOfficers = () => {
-    officers.sort(
-        (currentOfficerObject, nextOfficerObject) => {
-            const [currentOfficerObjectFirstName, currentOfficerObjectLastName] = currentOfficerObject.name.split(" ");
-            const [nextOfficerObjectFirstName, nextOfficerObjectLastName] = nextOfficerObject.name.split(" ");
-
-            if (currentOfficerObjectLastName < nextOfficerObjectLastName) { return -1; }
-            if (currentOfficerObjectLastName > nextOfficerObjectLastName) { return 1; }
-            return 0;
-        }
-    )
-    return officers.slice();
+    return sortArrayByLastName(officers).slice();
 };
 
 // Fetches a JSON string of officers data and then converts it to a JavaScript array.
