+++ styles/elements/criminal.css
@@ -152,5 +152,6 @@
 #listKeyContainer {
     display: flex;
     flex-wrap: wrap;
-    margin: 5rem 10% 0 10%;
+    justify-content: space-between;
+    margin: 5rem 5rem 0 5rem;
 }
